черновой коммит. разделение логики и представления модулей на первых двух экранах.
