+++ js/enter-next-level.js
@@ -1,7 +1,7 @@
 import {levels} from './data.js';
 import showScreen from './showscreen-function.js';
 import {getThreeCardsGameScreen} from './three-cards-game-screen.js';
-import {getTwoCardsGameScreen} from './two-cards-game-screen.js';
+import getTwoCardsGameScreen from './two-cards-game-screen.js';
 import {getOneCardGameScreen} from './one-card-game-screen.js';
 import getStatsScreenElement from './stats-screen.js';
 import {userState} from './utils.js';
