+++ js/main.js
@@ -1,3 +1,9 @@
-import GetMainScreenElement from './main-screen.js';
+import MainScreenView from './main-screen.js';
 import showScreen from './showscreen-function.js';
-showScreen(new GetMainScreenElement().element);
+import getGreetingScreen from './greeting-screen.js';
+
+const mainScreenView = new MainScreenView();
+mainScreenView.onClick = () => {
+  showScreen(getGreetingScreen().element);
+};
+showScreen(mainScreenView.element);
