+++ js/showscreen-function.js
@@ -1,5 +1,6 @@
 const bodyTag = document.querySelector(`body`);
 const showScreen = (el) => {
+  console.log(el);
   const oldScreen = bodyTag.querySelector(`.central`);
   bodyTag.replaceChild(el, oldScreen);
 };
